+++ lib/pychilizer/__init__.py
@@ -1,0 +1,3 @@
+"""pyChilizer module"""
+
+__author__ = 'Archilizer'
