+++ lib/pychilizer/database.py
@@ -1,0 +1,76 @@
+from pyrevit import revit, DB, script, forms, HOST_APP, coreutils
+from pyrevit.revit.db import query
+
+
+
+def get_sheet(some_number):
+    sheet_nr_filter = query.get_biparam_stringequals_filter({DB.BuiltInParameter.SHEET_NUMBER: str(some_number)})
+    found_sheet = DB.FilteredElementCollector(revit.doc) \
+        .OfCategory(DB.BuiltInCategory.OST_Sheets) \
+        .WherePasses(sheet_nr_filter) \
+        .WhereElementIsNotElementType().ToElements()
+
+    return found_sheet
+
+
+def get_view(some_name):
+    view_name_filter = query.get_biparam_stringequals_filter({DB.BuiltInParameter.VIEW_NAME: some_name})
+    found_view = DB.FilteredElementCollector(revit.doc) \
+        .OfCategory(DB.BuiltInCategory.OST_Views) \
+        .WherePasses(view_name_filter) \
+        .WhereElementIsNotElementType().ToElements()
+
+    return found_view
+
+
+def get_fam(some_name):
+    fam_name_filter = query.get_biparam_stringequals_filter({DB.BuiltInParameter.SYMBOL_FAMILY_NAME_PARAM: some_name})
+    found_fam = DB.FilteredElementCollector(revit.doc) \
+        .OfCategory(DB.BuiltInCategory.OST_GenericModel) \
+        .WherePasses(fam_name_filter) \
+        .WhereElementIsNotElementType().ToElements()
+
+    return found_fam
+
+
+def param_set_by_cat(cat):
+    # get all project type parameters of a given category
+    # can be used to gather parameters for UI selection
+    all_gm = DB.FilteredElementCollector(revit.doc).OfCategory(cat).WhereElementIsElementType().ToElements()
+    parameter_set = []
+    for gm in all_gm:
+        params = gm.Parameters
+        for p in params:
+            if p not in parameter_set and p.IsReadOnly == False:
+                parameter_set.append(p)
+    return parameter_set
+
+
+# 0001
+def create_sheet(sheet_num, sheet_name, titleblock):
+    sheet_num = str(sheet_num)
+
+    new_datasheet = DB.ViewSheet.Create(revit.doc, titleblock)
+    new_datasheet.Name = sheet_name
+
+    while get_sheet(sheet_num):
+        sheet_num = coreutils.increment_str(sheet_num, 1)
+    new_datasheet.SheetNumber = str(sheet_num)
+
+    return new_datasheet
+
+
+def set_anno_crop(v):
+    anno_crop = v.get_Parameter(DB.BuiltInParameter.VIEWER_ANNOTATION_CROP_ACTIVE)
+    anno_crop.Set(1)
+    return anno_crop
+
+
+def apply_vt(v, vt):
+    if vt:
+        v.ViewTemplateId = vt.Id
+    return
+
+
+def get_name(el):
+    return DB.Element.Name.__get__(el)
